+++ discovery-provider/compose/.env
@@ -12,3 +12,11 @@ audius_delegate_private_key=0x3873ed01bfb13621f9301487cc61326580614a5b99f3c33cf3
 audius_discprov_identity_service_url=http://audius-identity-service_identity-service_1:7000
 audius_discprov_user_metadata_service_url=http://cn1_creator-node_1:4000
 audius_discprov_blacklist_block_indexing_interval=5
+audius_eth_contracts_registry=0x26b9c682a87879Ef035c92E4F978c303eC69a84B
+audius_solana_track_listen_count_address=4ZN1TTLp76gGCZ7tzXLfV4cahuHaP6oHFuiVWQS3knTp
+audius_solana_endpoint=http://solana:8899
+audius_solana_signer_group_address=CkEr7jVopeNDSYGQYuJnHYE8b62hisw5qCGBzf9Pwxev
+audius_solana_signer_group_address=CkEr7jVopeNDSYGQYuJnHYE8b62hisw5qCGBzf9Pwxev
+audius_solana_signer_group_address=CkEr7jVopeNDSYGQYuJnHYE8b62hisw5qCGBzf9Pwxev
+audius_solana_signer_group_address=CkEr7jVopeNDSYGQYuJnHYE8b62hisw5qCGBzf9Pwxev
+audius_solana_signer_group_address=CkEr7jVopeNDSYGQYuJnHYE8b62hisw5qCGBzf9Pwxev
