+++ discovery-provider/src/app.py
@@ -23,7 +23,7 @@ from src.queries import queries, search, search_queries, health_check, notificat
     block_confirmation, skipped_transactions
 from src.api.v1 import api as api_v1
 from src.utils import helpers
-from src.utils.challenges import create_challenges
+from src.challenges.create_new_challenges import create_new_challenges
 from src.utils.multi_provider import MultiProvider
 from src.utils.session_manager import SessionManager
 from src.utils.config import config_files, shared_config, ConfigIni
@@ -302,7 +302,7 @@ def configure_flask(test_config, app, mode="app"):
     app.register_blueprint(api_v1.bp_full)
 
     # Create challenges
-    create_challenges(app.db_session_manager)
+    create_new_challenges(app.db_session_manager)
 
     return app
