+++ discovery-provider/src/challenges/challenges.json
@@ -2,7 +2,7 @@
 	{
 		"id": "profile_completion",
 		"type": "numeric",
-		"amount": 5 ,
+		"amount": "5",
 		"active": true,
 		"step_count": 7
 	}
